+++ src/test/java/de/uros/citlab/module/util/IModuleTest.java
@@ -6,14 +6,20 @@
 package de.uros.citlab.module.util;
 
 import eu.transkribus.interfaces.IModule;
+import org.junit.Assert;
+import org.junit.Test;
+
+import java.io.File;
+import java.util.List;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
 import static org.junit.Assert.fail;
 
 /**
- *
  * @author gundram
  */
 public class IModuleTest {
-
     public static void testUsage(IModule instance) {
         System.out.println("usage");
         String result = instance.usage();
@@ -37,6 +43,24 @@ public class IModuleTest {
             fail("Version does not fit to regex \"[0-9]+\\\\.[0-9]+\\\\.[0-9]+\".");
 
         }
+    }
+
+    @Test
+    public void testVersion() {
+        String softwareVersion = MetadataUtil.getSoftwareVersion();
+        File f = new File("pom.xml");
+        Pattern p = Pattern.compile(".*<version>(([0-9.]+).*(-SNAPSHOT)?)</version>.*");
+        List<String> strings = FileUtil.readLines(f);
+        strings.sort((o1, o2) -> Integer.compare(o1.indexOf('<'), o2.indexOf('<')));
+        for (String s : strings) {
+            Matcher matcher = p.matcher(s);
+            if (matcher.matches()) {
+                String group = matcher.group(1);
+                Assert.assertEquals("saved version does not fit to version in pom.xml", group, MetadataUtil.getSoftwareVersion());
+                break;
+            }
+        }
+
     }
 
     public static void testGetProvider(IModule instance) {