+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # Change Log
 
+# 2.1.1
+* bugfix training of HTR+ with base model
+
 # 2.1.0
 * faster TextAlignment and base technology moved to CITlabTextAlignment
 * HTR+ can be trained further - also with changed CharMap
 * bugfix load dictionary: now also possible for ";"-seperator with headlines
 * feature-request: logging in multi-process-use possible.
 * newest versions of all citlab and planet libs
+* bugfix fail-save Baseline2Polygon - 2 fallbacks for fails
 
 # 2.0.2
 * MOVE REPOSITORY WITHOUT HISTORY FROM Transkribs to CITlabRostock AND MAKE OPENSOURCE